Set up object oriented approach
Creates an object oriented approach to model the chessboard, chess
squares and chess pieces as "Chessboard", "ChessSquare" and
"ChessPiece".

Each object type has its own function constructor and sets its
properties. The initChessmaster function initializes the board and its
pieces and automatically displays it as a child of an element with the
id "chessboard".